backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pathlib
psycopg2-binary
passlib
pydantic